refactor: move authorize_interceptor directory
