+++ lib/app/core/di/module.dart
@@ -4,7 +4,7 @@ import 'package:dio_cookie_manager/dio_cookie_manager.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:injectable/injectable.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:ziggle/app/modules/auth/data/data_sources/authorize_interceptor.dart';
+import 'package:ziggle/app/modules/auth/data/data_sources/remote/authorize_interceptor.dart';
 
 @module
 abstract class Module {
